+++ src/scope.rs
@@ -12,7 +12,7 @@ pub struct Scope<'a, Recorder: ProfilerCommandRecorder> {
     pub scope: Option<GpuProfilerQuery>,
 }
 
-impl<'a, R: ProfilerCommandRecorder> Drop for Scope<'a, R> {
+impl<R: ProfilerCommandRecorder> Drop for Scope<'_, R> {
     #[inline]
     fn drop(&mut self) {
         if let Some(scope) = self.scope.take() {
@@ -30,7 +30,7 @@ pub struct OwningScope<'a, Recorder: ProfilerCommandRecorder> {
     pub scope: Option<GpuProfilerQuery>,
 }
 
-impl<'a, R: ProfilerCommandRecorder> Drop for OwningScope<'a, R> {
+impl<R: ProfilerCommandRecorder> Drop for OwningScope<'_, R> {
     #[inline]
     fn drop(&mut self) {
         if let Some(scope) = self.scope.take() {
@@ -51,7 +51,7 @@ pub struct ManualOwningScope<'a, Recorder: ProfilerCommandRecorder> {
     pub scope: Option<GpuProfilerQuery>,
 }
 
-impl<'a, R: ProfilerCommandRecorder> ManualOwningScope<'a, R> {
+impl<R: ProfilerCommandRecorder> ManualOwningScope<'_, R> {
     /// Ends the scope allowing the extraction of the owned [`ProfilerCommandRecorder`].
     #[track_caller]
     #[inline]
@@ -75,15 +75,11 @@ macro_rules! impl_scope_ext {
             #[must_use]
             #[track_caller]
             #[inline]
-            pub fn scope(
-                &mut self,
-                label: impl Into<String>,
-                device: &wgpu::Device,
-            ) -> Scope<'_, R> {
+            pub fn scope(&mut self, label: impl Into<String>) -> Scope<'_, R> {
                 let recorder: &mut R = &mut self.recorder;
                 let scope = self
                     .profiler
-                    .begin_query(label, recorder, device)
+                    .begin_query(label, recorder)
                     .with_parent(self.scope.as_ref());
                 Scope {
                     profiler: self.profiler,
@@ -99,23 +95,25 @@ macro_rules! impl_scope_ext {
             /// Ignores passed `wgpu::RenderPassDescriptor::timestamp_writes` and replaces it with
             /// `timestamp_writes` managed by `GpuProfiler` if profiling is enabled.
             ///
+            /// This also sets the `wgpu::RenderPassDescriptor::label` if it's `None` (default).
+            ///
             /// Note that in order to take measurements, this requires the [`wgpu::Features::TIMESTAMP_QUERY`] feature.
             /// [`wgpu::Features::TIMESTAMP_QUERY_INSIDE_ENCODERS`] & [`wgpu::Features::TIMESTAMP_QUERY_INSIDE_PASSES`] are not required.
             #[track_caller]
             pub fn scoped_render_pass(
                 &mut self,
                 label: impl Into<String>,
-                device: &wgpu::Device,
-                pass_descriptor: wgpu::RenderPassDescriptor,
-            ) -> OwningScope<wgpu::RenderPass> {
+                pass_descriptor: wgpu::RenderPassDescriptor<'_>,
+            ) -> OwningScope<'_, wgpu::RenderPass<'_>> {
                 let child_scope = self
                     .profiler
-                    .begin_pass_query(label, &mut self.recorder, device)
+                    .begin_pass_query(label, &mut self.recorder)
                     .with_parent(self.scope.as_ref());
                 let render_pass = self
                     .recorder
                     .begin_render_pass(&wgpu::RenderPassDescriptor {
                         timestamp_writes: child_scope.render_pass_timestamp_writes(),
+                        label: pass_descriptor.label.or(Some(&child_scope.label)),
                         ..pass_descriptor
                     });
 
@@ -134,14 +132,13 @@ macro_rules! impl_scope_ext {
             /// Note that in order to take measurements, this requires the [`wgpu::Features::TIMESTAMP_QUERY`] feature.
             /// [`wgpu::Features::TIMESTAMP_QUERY_INSIDE_ENCODERS`] & [`wgpu::Features::TIMESTAMP_QUERY_INSIDE_PASSES`] are not required.
             #[track_caller]
-            pub fn scoped_compute_pass<'b>(
-                &'b mut self,
+            pub fn scoped_compute_pass(
+                &mut self,
                 label: impl Into<String>,
-                device: &wgpu::Device,
-            ) -> OwningScope<'b, wgpu::ComputePass<'b>> {
+            ) -> OwningScope<'_, wgpu::ComputePass<'_>> {
                 let child_scope = self
                     .profiler
-                    .begin_pass_query(label, &mut self.recorder, device)
+                    .begin_pass_query(label, &mut self.recorder)
                     .with_parent(self.scope.as_ref());
 
                 let render_pass = self