+++ tests/src/mod.rs
@@ -1,6 +1,7 @@
 mod dropped_frame_handling;
 mod errors;
 mod interleaved_command_buffer;
+mod multiple_resolves_per_frame;
 mod nested_scopes;
 
 pub fn create_device(
@@ -9,7 +10,7 @@ pub fn create_device(
     async fn create_default_device_async(
         features: wgpu::Features,
     ) -> Result<(wgpu::Backend, wgpu::Device, wgpu::Queue), wgpu::RequestDeviceError> {
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
+        let instance = wgpu::Instance::new(&wgpu::InstanceDescriptor {
             backends: wgpu::Backends::PRIMARY, // Workaround for wgl having issues with parallel device destruction.
             ..Default::default()
         });
@@ -18,13 +19,10 @@ pub fn create_device(
             .await
             .unwrap();
         let (device, queue) = adapter
-            .request_device(
-                &wgpu::DeviceDescriptor {
-                    required_features: features,
-                    ..Default::default()
-                },
-                None,
-            )
+            .request_device(&wgpu::DeviceDescriptor {
+                required_features: features,
+                ..Default::default()
+            })
             .await?;
         Ok((adapter.get_info().backend, device, queue))
     }